+++ internal/handlers/academicSessions.go
