+++ internal/handlers/courses.go
